Refactor test organization: split `main_test.go` into dedicated packages for entries, filter, and parsing.
